Remove unsafe methods from test cases by using nix instead
This adds nix as a dev-dependency so that it can be used to replace the
direct libc calls within the test cases.
