+++ perf-event/tests/sampling/mmap.rs
@@ -1,31 +1,16 @@
+use super::Hex;
+use nix::unistd::SysconfVar;
 use perf_event::events::Software;
 use perf_event::samples::RecordEvent;
 use perf_event::Builder;
-use std::fmt;
-
-#[derive(Copy, Clone, Eq, PartialEq)]
-struct Hex<T>(T);
-
-impl<T: fmt::UpperHex> fmt::Display for Hex<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl<T: fmt::UpperHex> fmt::Debug for Hex<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("0x")?;
-        self.0.fmt(f)
-    }
-}
 
 #[test]
 fn record_executable_mmap() {
     // Use the pagesize as the mmap size so that the recorded length in the
     // record and the length we allocate are the same.
-    let pagesize = unsafe { libc::sysconf(libc::_SC_PAGESIZE) };
-    assert!(pagesize >= 0);
-    let pagesize = pagesize as usize;
+    let pagesize = nix::unistd::sysconf(SysconfVar::PAGE_SIZE)
+        .expect("Unable to get page size")
+        .expect("No page size returned") as usize;
 
     let mut sampler = Builder::new()
         .kind(Software::DUMMY)
@@ -57,6 +42,6 @@ fn record_executable_mmap() {
 
     assert_eq!(Hex(record.addr), Hex(mmap.as_ptr() as usize as _));
     assert_eq!(Hex(record.len), Hex(mmap.len() as _));
-    assert_eq!(Hex(record.pid), Hex(unsafe { libc::getpid() as _ }));
-    assert_eq!(Hex(record.tid), Hex(unsafe { libc::gettid() as _ }));
+    assert_eq!(Hex(record.pid), Hex(nix::unistd::getpid().as_raw() as _));
+    assert_eq!(Hex(record.tid), Hex(nix::unistd::gettid().as_raw() as _));
 }
